add folder existance guards to cleanup script

+++ cleanup.sh
@@ -1,3 +1,7 @@
 #/bin/bash
-rm /var/www/html/package.json
-rm -r /var/www/html/dist/*
+if [ -f /var/www/html/package.json ]; then
+    rm /var/www/html/package.json
+fi
+if compgen -G "/var/www/html/dist/*" > /dev/null; then
+    rm -r /var/www/html/dist/*
+fi
